✨ TWEAK: script command and remove network
